+++ GCPlus/Items/HiddenGoldItem.cs
@@ -1,6 +1,8 @@
-﻿using GoldenCoastPlusRevived.Modules;
+﻿using GoldenCoastPlusRevived.Buffs;
+using GoldenCoastPlusRevived.Modules;
 using RoR2;
 using UnityEngine;
+using UnityEngine.Networking;
 
 namespace GoldenCoastPlusRevived.Items
 {
@@ -16,13 +18,22 @@ namespace GoldenCoastPlusRevived.Items
         internal override GameObject modelPrefab => null;
         internal override Sprite iconSprite => GCPAssets.HiddenGoldBuffIcon;
         internal override ItemTier Tier => ItemTier.NoTier;
-        internal override ItemTag[] ItemTags => new ItemTag[] { ItemTag.BrotherBlacklist, ItemTag.CannotDuplicate };
+        internal override ItemTag[] ItemTags => new ItemTag[] { ItemTag.BrotherBlacklist, ItemTag.CannotDuplicate, ItemTag.WorldUnique };
         internal override bool hidden => true;
 
 
         internal override void AddHooks()
         {
             On.RoR2.CharacterMaster.GiveMoney += CharacterMaster_GiveMoney;
+            RoR2.CharacterBody.onBodyInventoryChangedGlobal += CharacterBody_onBodyInventoryChangedGlobal;
+        }
+
+        private void CharacterBody_onBodyInventoryChangedGlobal(CharacterBody body)
+        {
+            if (NetworkServer.active)
+            {
+                body.SetBuffCount(HiddenGoldBuff.BuffIndex, body.inventory.GetItemCount(HiddenGoldItem.ItemIndex));
+            }
         }
 
         private static void CharacterMaster_GiveMoney(On.RoR2.CharacterMaster.orig_GiveMoney orig, CharacterMaster self, uint amount)