+++ Thunderstore/CHANGELOG.md
@@ -1,8 +1,10 @@
 `1.2.2`
 
 - Fixed Boss attacks being out of order
-- Tweaked Laser Eye to scale with 2 (+2 per stack) enemies hit instead of 2500% (+2500% per stack) damage
 - Fixed things just straight up ignoring configs???? 
+- Added AutoConfigs and config syncing
+- Laser Eye now scales with 2 (+2 per stack) enemies hit instead of 2500% (+2500% per stack) damage
+- Laser Eye stacks no longer are timed
 
 `1.2.1`